requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
appdirs=1.4.4=pypi_0
audioread=3.0.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
bottleneck=1.3.5=py37h7deecbd_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=3.0.1=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.2.1=pypi_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.11.0=pypi_0
debugpy=1.6.6=pypi_0
decorator=5.1.1=pypi_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=pypi_0
filelock=3.10.0=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
fvcore=0.1.6=pypi_0
giflib=5.2.1=h5eee18b_1
gitdb=4.0.10=pypi_0
gitpython=3.1.30=pypi_0
google-auth=2.16.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grad-cam=1.4.6=pypi_0
grpcio=1.51.1=pypi_0
h5py=3.7.0=py37h737f45e_0
hdf5=1.10.6=h3ffc7dd_1
huggingface-hub=0.13.2=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.10=pypi_0
ipykernel=6.16.2=pypi_0
ipython=7.34.0=pypi_0
jedi=0.18.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9b=h024ee3a_2
jupyter-client=7.4.9=pypi_0
jupyter-core=4.12.0=pypi_0
kiwisolver=1.4.4=pypi_0
lcms2=2.12=h3be6417_0
libedit=3.1.20221030=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.37=hbc83047_0
librosa=0.9.2=pypi_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=h85742a9_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
llvmlite=0.39.1=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numba=0.56.4=pypi_0
numexpr=2.8.4=py37he184ba9_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.0.2u=h7b6447c_0
packaging=23.0=pypi_0
pandas=1.3.4=py37h8c16a72_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.3.0=py37hace64e9_1
pip=22.3.1=py37h06a4308_0
pooch=1.6.0=pypi_0
portalocker=2.7.0=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pygments=2.14.0=pypi_0
pyparsing=3.0.9=pypi_0
python=3.7.0=h6e4f718_3
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.7.1=py3.7_cuda11.0.221_cudnn8.0.5_0
pytz=2022.7.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.1=pypi_0
readline=7.0=h7b6447c_5
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
resampy=0.4.2=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
sentry-sdk=1.14.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.6.3=py37h06a4308_0
simplejson=3.18.3=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
soundfile=0.11.0=pypi_0
spyder-kernels=2.4.2=pypi_0
sqlite=3.33.0=h62c20be_0
tabulate=0.9.0=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.2.0=pypi_0
threadpoolctl=3.1.0=pypi_0
timm=0.6.12=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu110
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.9.0=pypi_0
ttach=0.0.3=pypi_0
typing_extensions=4.4.0=py37h06a4308_0
urllib3=1.26.14=pypi_0
wandb=0.13.9=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wurlitzer=3.0.3=pypi_0
xz=5.2.10=h5eee18b_1
yacs=0.1.8=pypi_0
zipp=3.12.1=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=haebb681_0
